Introduce fastboot.efi
fastboot.efi is an implementation of a subset of the fastboot protocol
as a standalone EFI application, for the purpose of enabling workflows
using "fastboot boot" for e.g. rapid kernel development.

It enables "fastboot boot" of abootimg v2 files, as well as PE32+ (EFI)
applications, making it possible to boot either abootimg or e.g. UKI
images.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>