abootimg: Provide user feedback for errors
Most users attempts to pass a compressed Image in the abootimg, which
isn't supported by UEFI. But as we don't provide suitable error messages
there's no way to know the cause.

So, introduce a "with_context" method for Result, in order to pass
context with the various failures and propagate this to the FAIL message
returned to the user.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>