abootimg: Special case the abootimg v0 support
When the user invoke "fastboot boot foo.uki", the official fastboot tool
will wrap "foo.uki" in a abootimg header version 0; with ramdisk and
second size of 0.

Add special handling for this case, to allow booting UKIs this way, but
push the user to use version 2 if there's other content in the image, as
this likely indicates that they are trying to boot a abootimg with DTB
appended to the kernel.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>